feat: add watcher support to DID creation and updates
- Introduced a new `--watcher` option in the CLI for adding watcher URLs.
- Updated `handleCreate` and `handleUpdate` functions to handle watchers.
- Enhanced interfaces to include `watchers` property.
- Added tests for creating and updating DIDs with watchers, ensuring persistence and functionality.
- Updated existing logs and tests to reflect the new versioning scheme (1.0).
