+++ src/constants.ts
@@ -1,7 +1,5 @@
 export const PLACEHOLDER = "{SCID}";
 export const METHOD = "webvh";
-export const VERSION = "0.5";
-export const PROTOCOL = `did:${METHOD}:${VERSION}`;
 export const BASE_CONTEXT = [
   "https://www.w3.org/ns/did/v1",
   "https://w3id.org/security/multikey/v1"
