+++ src/routes/.well-known/did.jsonl
@@ -1,1 +1,1 @@
-{"versionId":"1-QmW9zHgCsAx5KgY2qPxsNiabVhzm6vxpqRyYrCFTbt5sU8","versionTime":"2025-01-15T06:27:42Z","parameters":{"method":"did:webvh:0.5","scid":"QmcvPyyDtWEz8FRDwE1j21YrZD2cCkNJzbUtewCXVbrB9d","updateKeys":["z6MkvDJLghkVAzigMw3jZGQ5qMTDnb3iprMqDiyR4P6GVTpQ"],"portable":false,"nextKeyHashes":[],"witnesses":[],"witnessThreshold":0,"deactivated":false},"state":{"@context":["https://www.w3.org/ns/did/v1","https://w3id.org/security/multikey/v1"],"id":"did:webvh:QmcvPyyDtWEz8FRDwE1j21YrZD2cCkNJzbUtewCXVbrB9d:localhost%3A8000","controller":"did:webvh:QmcvPyyDtWEz8FRDwE1j21YrZD2cCkNJzbUtewCXVbrB9d:localhost%3A8000","assertionMethod":["did:webvh:QmcvPyyDtWEz8FRDwE1j21YrZD2cCkNJzbUtewCXVbrB9d:localhost%3A8000#4P6GVTpQ"],"verificationMethod":[{"id":"did:webvh:QmcvPyyDtWEz8FRDwE1j21YrZD2cCkNJzbUtewCXVbrB9d:localhost%3A8000#4P6GVTpQ","controller":"did:webvh:QmcvPyyDtWEz8FRDwE1j21YrZD2cCkNJzbUtewCXVbrB9d:localhost%3A8000","type":"Multikey","publicKeyMultibase":"z6MkvDJLghkVAzigMw3jZGQ5qMTDnb3iprMqDiyR4P6GVTpQ"}]},"proof":[{"type":"DataIntegrityProof","cryptosuite":"eddsa-jcs-2022","verificationMethod":"did:key:z6MkvDJLghkVAzigMw3jZGQ5qMTDnb3iprMqDiyR4P6GVTpQ#z6MkvDJLghkVAzigMw3jZGQ5qMTDnb3iprMqDiyR4P6GVTpQ","created":"2025-01-15T06:27:42Z","proofPurpose":"assertionMethod","proofValue":"z3Go6ZqYWoUQ7myyCaP6hG4Xd3iSGv4swnvh3qcHrfnc7iHfyEd46vGXnRWzCSm7VQbqv7Mo2RNyuiEFkip9ko5Fd"}]}
+{"versionId":"1-QmW9zHgCsAx5KgY2qPxsNiabVhzm6vxpqRyYrCFTbt5sU8","versionTime":"2025-01-15T06:27:42Z","parameters":{"method":"did:webvh:1.0","scid":"QmcvPyyDtWEz8FRDwE1j21YrZD2cCkNJzbUtewCXVbrB9d","updateKeys":["z6MkvDJLghkVAzigMw3jZGQ5qMTDnb3iprMqDiyR4P6GVTpQ"],"portable":false,"nextKeyHashes":[],"witnesses":[],"witnessThreshold":0,"deactivated":false},"state":{"@context":["https://www.w3.org/ns/did/v1","https://w3id.org/security/multikey/v1"],"id":"did:webvh:QmcvPyyDtWEz8FRDwE1j21YrZD2cCkNJzbUtewCXVbrB9d:localhost%3A8000","controller":"did:webvh:QmcvPyyDtWEz8FRDwE1j21YrZD2cCkNJzbUtewCXVbrB9d:localhost%3A8000","assertionMethod":["did:webvh:QmcvPyyDtWEz8FRDwE1j21YrZD2cCkNJzbUtewCXVbrB9d:localhost%3A8000#4P6GVTpQ"],"verificationMethod":[{"id":"did:webvh:QmcvPyyDtWEz8FRDwE1j21YrZD2cCkNJzbUtewCXVbrB9d:localhost%3A8000#4P6GVTpQ","controller":"did:webvh:QmcvPyyDtWEz8FRDwE1j21YrZD2cCkNJzbUtewCXVbrB9d:localhost%3A8000","type":"Multikey","publicKeyMultibase":"z6MkvDJLghkVAzigMw3jZGQ5qMTDnb3iprMqDiyR4P6GVTpQ"}]},"proof":[{"type":"DataIntegrityProof","cryptosuite":"eddsa-jcs-2022","verificationMethod":"did:key:z6MkvDJLghkVAzigMw3jZGQ5qMTDnb3iprMqDiyR4P6GVTpQ#z6MkvDJLghkVAzigMw3jZGQ5qMTDnb3iprMqDiyR4P6GVTpQ","created":"2025-01-15T06:27:42Z","proofPurpose":"assertionMethod","proofValue":"z3Go6ZqYWoUQ7myyCaP6hG4Xd3iSGv4swnvh3qcHrfnc7iHfyEd46vGXnRWzCSm7VQbqv7Mo2RNyuiEFkip9ko5Fd"}]}
