--- test/fixtures/not-authorized.log
@@ -1,2 +1,0 @@
-{"versionId":"1-QmbxtK5M3NqxwqdSetjVQdN1JT5nHnXPizM58Hqcf21GSr","versionTime":"2025-02-10T23:29:57Z","parameters":{"method":"did:webvh:0.5","scid":"QmZqUStGokTHhsDaR6bgzg5sF15cyv8qnAnjfRqAwhBHZQ","updateKeys":["z6MkpGJzzXoh3yddHvnw8EB6btRvpwGrRrqJYZcuDW3ty2b2"],"portable":false,"nextKeyHashes":[],"deactivated":false},"state":{"@context":["https://www.w3.org/ns/did/v1","https://w3id.org/security/multikey/v1"],"id":"did:webvh:QmZqUStGokTHhsDaR6bgzg5sF15cyv8qnAnjfRqAwhBHZQ:example.com","controller":"did:webvh:QmZqUStGokTHhsDaR6bgzg5sF15cyv8qnAnjfRqAwhBHZQ:example.com","assertionMethod":["did:webvh:QmZqUStGokTHhsDaR6bgzg5sF15cyv8qnAnjfRqAwhBHZQ:example.com#DW3ty2b2","did:webvh:QmZqUStGokTHhsDaR6bgzg5sF15cyv8qnAnjfRqAwhBHZQ:example.com#FpReMRf7"],"verificationMethod":[{"id":"did:webvh:QmZqUStGokTHhsDaR6bgzg5sF15cyv8qnAnjfRqAwhBHZQ:example.com#DW3ty2b2","controller":"did:webvh:QmZqUStGokTHhsDaR6bgzg5sF15cyv8qnAnjfRqAwhBHZQ:example.com","type":"Multikey","publicKeyMultibase":"z6MkpGJzzXoh3yddHvnw8EB6btRvpwGrRrqJYZcuDW3ty2b2"},{"id":"did:webvh:QmZqUStGokTHhsDaR6bgzg5sF15cyv8qnAnjfRqAwhBHZQ:example.com#FpReMRf7","controller":"did:webvh:QmZqUStGokTHhsDaR6bgzg5sF15cyv8qnAnjfRqAwhBHZQ:example.com","type":"Multikey","publicKeyMultibase":"z6MkubmmkTKBgrJUPM3GAc8pXvFXrwcjWDKL2yn1FpReMRf7"}]},"proof":[{"type":"DataIntegrityProof","cryptosuite":"eddsa-jcs-2022","verificationMethod":"did:key:z6MkpGJzzXoh3yddHvnw8EB6btRvpwGrRrqJYZcuDW3ty2b2#z6MkpGJzzXoh3yddHvnw8EB6btRvpwGrRrqJYZcuDW3ty2b2","created":"2025-02-10T23:29:57Z","proofPurpose":"assertionMethod","proofValue":"z2sSpRVngknJ5ZfTPnNjw5BbFYoVZv2PJvvDMB1n46zXvugHHC6uPaK13P4YPczhamc1jfFtPbEFfnPMRWvrjMaFK"}]}
-{"versionId":"2-QmQdmvrWatDsCfH2FSbGXpgweSwByFxjG5kHub9m8RVT2b","versionTime":"2025-02-10T23:29:57Z","parameters":{},"state":{"@context":["https://www.w3.org/ns/did/v1","https://w3id.org/security/multikey/v1"],"id":"did:webvh:QmZqUStGokTHhsDaR6bgzg5sF15cyv8qnAnjfRqAwhBHZQ:example.com","controller":["did:webvh:QmZqUStGokTHhsDaR6bgzg5sF15cyv8qnAnjfRqAwhBHZQ:example.com"],"assertionMethod":["did:webvh:QmZqUStGokTHhsDaR6bgzg5sF15cyv8qnAnjfRqAwhBHZQ:example.com#WqEzhjrq","did:webvh:QmZqUStGokTHhsDaR6bgzg5sF15cyv8qnAnjfRqAwhBHZQ:example.com#TjbBnqi3"],"verificationMethod":[{"id":"did:webvh:QmZqUStGokTHhsDaR6bgzg5sF15cyv8qnAnjfRqAwhBHZQ:example.com#WqEzhjrq","controller":"did:webvh:QmZqUStGokTHhsDaR6bgzg5sF15cyv8qnAnjfRqAwhBHZQ:example.com","type":"Multikey","publicKeyMultibase":"z6MktcKXiMUPUA8HdTVsdMQaz1zeRYe7SA12pJRuWqEzhjrq"},{"id":"did:webvh:QmZqUStGokTHhsDaR6bgzg5sF15cyv8qnAnjfRqAwhBHZQ:example.com#TjbBnqi3","controller":"did:webvh:QmZqUStGokTHhsDaR6bgzg5sF15cyv8qnAnjfRqAwhBHZQ:example.com","type":"Multikey","publicKeyMultibase":"z6MkhU744oqpCWW3p7qvRh6fBB3SsMMpY5XWvFdhTjbBnqi3"}]},"proof":[{"type":"DataIntegrityProof","cryptosuite":"eddsa-jcs-2022","verificationMethod":"did:key:z6MkubmmkTKBgrJUPM3GAc8pXvFXrwcjWDKL2yn1FpReMRf7#z6MkubmmkTKBgrJUPM3GAc8pXvFXrwcjWDKL2yn1FpReMRf7","created":"2025-02-10T23:29:57Z","proofPurpose":"assertionMethod","proofValue":"z2UPBL9SnP1Z9vSryu7vbH2RcCY8EJGSkGkCgLh6gzwqqCTFRiVqfwu1bjTuNxDC4iqrW3NSgFH5mPunkGBR4jqst"}]}
